requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-retry==2.9.1
aiosignal==1.3.2
alembic==1.14.0
amqp==5.3.1
aniso8601==9.0.1
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==25.1.0
babel==2.17.0
bcrypt==4.2.1
beautifulsoup4==4.13.3
billiard==4.2.1
bleach==6.2.0
blinker==1.9.0
Brotli==1.1.0
celery==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cssselect2==0.8.0
cycler==0.12.1
debugpy==1.8.12
decorator==5.2.1
defusedxml==0.7.1
dnspython==2.6.1
et_xmlfile==2.0.0
executing==2.2.0
fastjsonschema==2.21.1
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.0
Flask-MySQL==1.6.0
Flask-MySQLdb==2.0.0
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==9.0.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.8
kombu==5.5.3
Mako==1.3.8
MarkupSafe==2.1.5
matplotlib==3.10.1
matplotlib-inline==0.1.7
mistune==3.1.2
multidict==6.4.3
mysqlclient==2.2.6
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.3.2
notebook_shim==0.2.4
numpy==2.2.5
openpyxl==3.1.5
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
phonenumbers==9.0.3
pillow==11.1.0
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.1
psutil==7.0.0
pure_eval==0.2.3
pycparser==2.22
pydyf==0.11.0
Pygments==2.19.1
PyJWT==2.9.0
pymango==0.1.1
pymongo==4.11.3
PyMySQL==1.1.1
pyparsing==3.2.3
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
pytz==2024.2
pywin32==308
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.2.1
redis==5.2.1
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.23.1
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==75.8.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
stack-data==0.6.3
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tornado==6.4.2
traitlets==5.14.3
twilio==9.5.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.2
uri-template==1.3.0
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==64.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.1
yarl==1.19.0
zopfli==0.2.3.post1
